refactor yaml reader function into utisl
